// plugins/woocommerce/client/blocks/assets/js/blocks/product-search/constants.ts
export const SEARCH_BLOCK_NAME = 'core/search';
export const SEARCH_VARIATION_NAME = 'woocommerce/product-search';

export enum PositionOptions {
	OUTSIDE = 'button-outside',
	INSIDE = 'button-inside',
	NO_BUTTON = 'no-button',
	BUTTON_ONLY = 'button-only',
	INPUT_AND_BUTTON = 'input-and-button',
}